+++ src/Providers/AbstractProvider.php
@@ -83,6 +83,16 @@ abstract class AbstractProvider
     abstract public function getFormat();
 
     /**
+     * Get source attribution
+     *
+     * @return string
+     */
+    public function getSource()
+    {
+        return $this->getShortName();
+    }
+
+    /**
      * Get http client options based on current client
      *
      * @return array
@@ -115,11 +125,24 @@ abstract class AbstractProvider
 
         $listings = $this->getRawListings($payload);
 
+        return $this->getJobsCollectionFromListings($listings);
+    }
+
+    /**
+     * Create and get collection of jobs from given listings
+     *
+     * @param  array $listings
+     *
+     * @return Collection
+     */
+    private function getJobsCollectionFromListings($listings)
+    {
         $collection = new Collection;
 
         array_map(function ($item) use ($collection) {
             $job = $this->createJobObject($item);
-            $job->setQuery($this->keyword);
+            $job->setQuery($this->keyword)
+                ->setSource($this->getSource());
             $collection->add($job);
         }, $listings);
 
@@ -158,6 +181,22 @@ abstract class AbstractProvider
         }
 
         return (array) $payload;
+    }
+
+    /**
+     * Get short name of a given or current class
+     *
+     * @return string
+     */
+    private function getShortName($object = null)
+    {
+        if (is_null($object)) {
+            $object = $this;
+        }
+
+        $ref = new \ReflectionClass(get_class($object));
+
+        return $ref->getShortName();
     }
 
     /**